+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Change color of indicator and borders in high contrast mode (light scheme) for radio and checkbox components (#645)
+- Change color of button loader in high contrast light mode (#437)
 
 ## [0.15.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.14.0...0.15.0) - 2025-05-28